- fix single row interator for builder client - moved logger package to plugin and removed notify
